refactor(lessons/repositories): alterados imports e exports
