+++ src/modules/lessons/infra/typeorm/repositories/Videos.repository.ts
@@ -1,12 +1,12 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { CreateVideoDto } from '@lessons:dtos/CreateVideo.dto';
-import Video from '@lessons:entities/Video';
-import IVideosRepository from '@lessons:irepos/IVideos.repository';
+import { CreateVideoDto } from '@lessons:dtos/videos/CreateVideo.dto';
+import { Video } from '@lessons:entities/Video';
+import { IVideosRepository } from '@lessons:irepos/IVideos.repository';
 
 // ---------------------------------------------------------------------------------------------- //
 
-class VideosRepository implements IVideosRepository {
+export class VideosRepository implements IVideosRepository {
   private repository: Repository<Video>;
 
   constructor() {
@@ -47,7 +47,3 @@ class VideosRepository implements IVideosRepository {
     return this.repository.findOne(id);
   }
 }
-
-// ---------------------------------------------------------------------------------------------- //
-
-export default VideosRepository;
